Lotta stuff
Most recent it `tools/infiltration.js` - Oh, yeah!  Easy billions of
money even on some restricted nodes.  166k rep for a faction by
taking on MegaCorp.  From a script found on Reddit, and I added an
'auto' option and 'faction' option to keep replaying and pick money
by default or a faction of your choice for reward.

Second, I added all my simple corp scripts and some in samples,
early-corp.js being very nice.

Also borrowed sample/contractor.js for auto finding and solving
contracts.

Also simple singularity and gang scripts.

Tweaks to start-hgw.js and some other scripts.  New mega-hack script
for when your hacking skill is just too high, like when I had the
weaken time on MegaCorp being just 33ms and 1 hack thread taking
almost 90%.   It uses zombies for each HGW and runs them at the same
time.  About 4ms extra time for each.

+++ main/start-grind.js
@@ -1,16 +1,45 @@
+// TODO:
+// 1. Fix --monitor, which should not kill anything, but should
+
+// args: --host <host,host,host> - add hosts to default of 'grinder', 'grinder-[0-24]'
+//       --gb <gb> - ram to reserve on 'home' only
+
 import { createTable } from './lib.js'
 
 /** @param {NS} ns */
 export async function main(ns) {
+  let args = [...ns.args]
+  let options = {}
+  const stripOption = (optionName, defaultValue) => {
+    for (let i = 0; i < args.length - 1; i++) {
+      if (args[i] === '--' + optionName) {
+        if (typeof(defaultValue) === 'undefined') {
+          // TODO: fix this, not working for --monitor
+          options[optionName] = true
+          args = args.slice(0, i).concat(args.slice(i + 1));
+        } else {
+          options[optionName] = args[i + 1];
+          args = args.slice(0, i).concat(args.slice(i + 2));
+        }
+        return options[optionName] }
+    }
+    options[optionName] = defaultValue
+    return defaultValue
+  }
+  // let host = stripOption('host', ns.getHostname())
+  let reserveGb = stripOption('gb', 32)
+  let monitor = stripOption('monitor')
+
   ns.disableLog('ALL')
   let hosts = ['grinder']
-  let targets = ['joesguns', 'foodnstuff', 'n00dles', 'sigma-cosmetics', 'nectar-net',
-    'hong-fang-tea', 'harakiri-sushi', 'neo-net', 'zer0', 'max-hardware', 'iron-gym']
+  // let targets = ['joesguns', 'foodnstuff', 'n00dles', 'sigma-cosmetics', 'nectar-net', 'hong-fang-tea', 'harakiri-sushi', 'neo-net', 'zer0', 'max-hardware', 'iron-gym']
+  let targets = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25].map(x => 'joesguns')
+
   for (let i = 0; i < targets.length - 1; i++) {
     hosts.push(`grinder-${i}`)
   }
-  if (ns.args[0]) {
-    hosts = ns.args[0].split(',').concat(hosts)
+  if (args[0]) {
+    hosts = args[0].split(',').concat(hosts)
   }
   hosts = hosts.filter(x => { try { 
     return ns.getServerMaxRam(x) > 0
@@ -20,34 +49,51 @@ export async function main(ns) {
     ns.tprint('ERROR: No hosts')
     return
   }
-  targets = targets.slice(0, hosts.length)
 
-  let grinderHost = ns.getHostname()
-  let killPids = ns.ps(grinderHost).filter(x => x.filename.indexOf("test-grind") >= 0).map(x => x.pid)
-  for (let i = 0; i < killPids.length; i++) {
-    ns.kill(killPids[i], grinderHost)
-    await ns.sleep(50)
+  // targets = targets.slice(0, hosts.length)
+  targets = hosts.map(x => 'joesguns') // always the best, works for 20 grinding servers for me
+
+  if (!monitor) { // TODO: Fix this, probably just not getting the value right above
+    let grinderHost = ns.getHostname()
+    let killPids = ns.ps(grinderHost).filter(x => x.filename.indexOf("test-grind") >= 0).map(x => x.pid)
+    for (let i = 0; i < killPids.length; i++) {
+      ns.kill(killPids[i], grinderHost)
+      await ns.sleep(50)
+    }
+    if (killPids.length) ns.tprint(`INFO: grind - killed ${killPids.length} scripts on ${grinderHost}`)
+
+    let pids = []
+    for (let i = 0; i < targets.length; i++) {
+      const host = hosts[i]
+      const target = targets[i]
+      ns.print(`Starting to target ${target} on ${host}`)
+      let args = [host, target]
+      if (host === 'home' && reserveGb) {
+        args.push(reserveGb)
+      }
+      pids.push(ns.run('/test/test-grind.js', 1, ...args))
+      await ns.sleep(200)
+    }
   }
-  if (killPids.length) ns.tprint(`INFO: grind - killed ${killPids.length} scripts on ${grinderHost}`)
-
-  let pids = []
-  for (let i = 0; i < targets.length; i++) {
-    const host = hosts[i]
-    const target = targets[i]
-    ns.print(`Starting to target ${target} on ${host}`)
-    pids.push(ns.run('/test/test-grind.js', 1, host, target))
-    await ns.sleep(200)
-  }
-
 
   while (true) {
-    let results = targets.map((target, index) => {
-      let host = hosts[index]
-      let income = ns.getScriptIncome('/remote/hack-loop.js', host, target, 0) || 0
-      let expGain = ns.getScriptExpGain('/remote/hack-loop.js', host, target, 0) || 0
-      return { target, host, income, expGain }
-    })
-    results.push({
+    let results = []
+    for (let i = 0; i < targets.length; i++) {
+      let target = targets[i];
+      let host = hosts[i]
+      try {
+        let income = ns.getScriptIncome('/remote/hack-loop.js', host, target, 0) || 0
+        let expGain = ns.getScriptExpGain('/remote/hack-loop.js', host, target, 0) || 0
+        results.push({ target, host, income, expGain })
+        await ns.sleep(100)
+      } catch (err) {
+        // maybe server was deleted?
+        targets.splice(i, 1)
+        hosts.splice(i, 1)
+        i--
+      }
+    }
+    results.unshift({
       host: '', target: 'total',
       income: results.reduce((p, c) => p + c.income, 0) || 0,
       expGain: results.reduce((p, c) => p + c.expGain, 0) || 0
@@ -60,6 +106,6 @@ export async function main(ns) {
     })), { align: { host: 'left', target: 'left' }})
     ns.clearLog()
     ns.print(table.join('\n'))
-    await ns.sleep(15000)
+    await ns.sleep(10000)
   }
 }
